move code generators from service_generator to service_generator_utils
